+++ src/quokka_editor_back/models/document.py
@@ -1,16 +1,8 @@
-from enum import StrEnum
-
 from tortoise import fields, models
 
 from quokka_editor_back.models.operation import Operation
 from quokka_editor_back.models.project import Project
 from quokka_editor_back.models.user import User
-
-
-class ShareRole(StrEnum):
-    READ = "READ"
-    COMMENT = "COMMENT"
-    EDIT = "EDIT"
 
 
 class Document(models.Model):
@@ -36,8 +28,6 @@ class Document(models.Model):
         on_delete=fields.CASCADE,
     )
     last_revision = fields.BigIntField(default=0)
-    shared_role = fields.CharEnumField(ShareRole, default=ShareRole.READ)
-    shared_by_link = fields.BooleanField(default=False)
 
 
 class DocumentTemplate(models.Model):
