+++ src/quokka_editor_back/models/project.py
@@ -1,6 +1,14 @@
+from enum import StrEnum
+
 from tortoise import fields, models
 
 from quokka_editor_back.models.user import User
+
+
+class ShareRole(StrEnum):
+    READ = "READ"
+    COMMENT = "COMMENT"
+    EDIT = "EDIT"
 
 
 class Project(models.Model):
@@ -13,3 +21,5 @@ class Project(models.Model):
         on_delete=fields.CASCADE,
     )
     images = fields.JSONField(null=True)
+    shared_role = fields.CharEnumField(ShareRole, default=ShareRole.READ)
+    shared_by_link = fields.BooleanField(default=False)
