+++ src/quokka_editor_back/schema/document.py
@@ -2,8 +2,6 @@ from uuid import UUID
 
 from pydantic import BaseModel, Field
 
-from quokka_editor_back.models.document import ShareRole
-
 
 class DocumentCreatePayload(BaseModel):
     template_id: UUID | None = None
@@ -13,8 +11,3 @@ class DocumentCreatePayload(BaseModel):
 class DocumentUpdatePayload(BaseModel):
     title: str | None = Field(None, max_length=255)
     content: list[str] | None = Field(None)
-
-
-class ShareInput(BaseModel):
-    shared_role: ShareRole
-    shared_by_link: bool
