+++ src/quokka_editor_back/schema/project.py
@@ -1,6 +1,8 @@
 from fastapi import File, UploadFile
 from pydantic import BaseModel, Field
 
+from quokka_editor_back.models.project import ShareRole
+
 
 class ProjectCreatePayload(BaseModel):
     title: str = Field(max_length=255)
@@ -12,3 +14,8 @@ class ProjectUpdatePayload(BaseModel):
 
 class ProjectImagePayload(BaseModel):
     file: UploadFile = File(...)
+
+
+class ShareInput(BaseModel):
+    shared_role: ShareRole
+    shared_by_link: bool
